Move SocialLink component to atoms
