+++ components/organisms/footer.tsx
@@ -1,10 +1,8 @@
 import {
-	Link,
 	Box,
 	Flex,
 	Text,
 	Center,
-	LinkProps,
 } from '@chakra-ui/core';
 import NextLink from 'next/link';
 import {
@@ -12,35 +10,8 @@ import {
 	TwitterIcon,
 	EmailIcon,
 } from '../../theme/icons';
-import logOutboundLink from '../../utils/log-outbound-link';
 import Logo from '../atoms/logo';
-
-type SocialLinkProps = {
-	name: string;
-} & LinkProps;
-
-const SocialLink = ({ name, children, ...props }: SocialLinkProps) => {
-	return (
-		<Link
-			isExternal
-			_focus={{
-				outline: 'none',
-				color: 'white',
-			}}
-			_hover={{ color: 'white' }}
-			aria-label={name}
-			flex="1"
-			py={{ base: 8, md: 4 }}
-			transition="all 0.2s ease-in"
-			onClick={() => logOutboundLink(`Footer > ${name}`)}
-			{...props}
-		>
-			<Center>
-				{children}
-			</Center>
-		</Link>
-	);
-};
+import SocialLink from '../atoms/social-link';
 
 const Footer = () => {
 	return (
